+++ lib/flutter_storybook.dart
@@ -8,6 +8,7 @@ export 'src/knobs/string_knob.dart';
 export 'src/plugins/directionality.dart';
 export 'src/plugins/plugin.dart';
 export 'src/plugins/plugin_panel.dart';
+export 'src/plugins/localization.dart';
 export 'src/tests/devices.dart';
 export 'src/tests/storybook_tester.dart';
 export 'src/story.dart';